requirements.txt
praw>=7.7.0
python-docx>=1.1.0
openai>=1.30.0
apscheduler>=3.10.0
python-dotenv>=1.0.0
rich>=13.0.0
click>=8.1.0
sqlite-utils>=3.35.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
schedule>=1.2.0
